+++ src/components/Experiments/ExperimentsList/ExperimentsList.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/colors';
-@import '../../styles/vars';
+@import '../../../styles/colors';
+@import '../../../styles/vars';
 
 .expList {
   display: flex;
